Modulize the frontend code
