+++ python/hal9/login/frontend/src/types.ts
@@ -1,0 +1,13 @@
+export interface LoginInfo {
+  user: string;
+  photo: string;
+}
+
+export function isLoginInfo(obj: any): obj is LoginInfo {
+  return (
+    "user" in obj &&
+    typeof obj["user"] === "string" &&
+    "photo" in obj &&
+    typeof obj["photo"] === "string"
+  );
+}
